requirements.txt
streamlit
dotenv
cohere
langchain
chromadb
protobuf~=3.20
langchain-chroma
langchain-cohere
pypdf
pytesseract
Pillow